feature/SD-53: Docker-compose is working
