+++ .dockerignore
@@ -1,0 +1,5 @@
+__pycache__
+venv
+migrations
+.gitignore
+.idea
