+++ .flaskenv
@@ -1,2 +1,2 @@
-FLASK_APP=app
+FLASK_APP=core/__init__.py
 FLASK_DEBUG=1
