+++ core/models/__init__.py
@@ -1,0 +1,9 @@
+from core.models.project_type import ProjectTypeModel
+from core.models.user import UserModel
+from core.models.attendance import AttendanceModel
+from core.models.channel import ChannelModel
+from core.models.occupancy import OccupancyModel
+from core.models.position import PositionModel
+from core.models.project import ProjectModel
+from core.models.projects_positions_users import ProjectsPositionsUsers
+from core.models.role import RoleModel
