+++ core/models/project_type.py
@@ -1,10 +1,10 @@
 from datetime import datetime
 
-from db import db
+from core.db import db
 
 
-class RoleModel(db.Model):
-    __tablename__ = "roles"
+class ProjectTypeModel(db.Model):
+    __tablename__ = "project_types"
 
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String, unique=True)
@@ -14,7 +14,11 @@ class RoleModel(db.Model):
     created_at = db.Column(db.DateTime, default=datetime.utcnow)
     updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
 
-    users = db.relationship("UserModel", back_populates="role", lazy="dynamic", cascade="all, delete-orphan")
+    projects = db.relationship(
+        "ProjectModel",
+        back_populates="project_type",
+        cascade="all, delete-orphan"
+    )
 
     def __repr__(self):
-        return f"<Role: {self.name}>"
+        return f"<Project_type: {self.name}>"
