+++ core/models/projects_positions_users.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from db import db
+from core.db import db
 
 
 class ProjectsPositionsUsers(db.Model):
