+++ core/resources/__init__.py
@@ -1,0 +1,2 @@
+from core.resources.project import *
+from core.resources.project_type import *
