--- models/__init__.py
@@ -1,9 +1,0 @@
-from models.user import UserModel
-from models.role import RoleModel
-from models.occupancy import OccupancyModel
-from models.projects_positions_users import ProjectsPositionsUsers
-from models.project import ProjectModel
-from models.position import PositionModel
-from models.project_type import ProjectTypeModel
-from models.channel import ChannelModel
-from models.attendance import AttendanceModel
